refactor: move _app to features
